Phase F-1 wave 2: extract Worker + Group mixins (#114)
- `_ops/workers.py` — `_WorkerOpsMixin(_StoreCore)`: register_worker,
  reissue_credential, verify_worker_credential, read_worker,
  list_workers + the credential helpers and the class-level
  `_PASSWORD_HASHER` / `_UNKNOWN_WORKER_DUMMY_HASH` (90 SLOC).
- `_ops/groups.py` — `_GroupOpsMixin(_StoreCore)`: register_group,
  add_to_group, remove_from_group, delete_group, read_group,
  list_groups, resolve_worker_in_group, _require_no_cycle_after
  (142 SLOC).
- `_StoreCore` gains a `resolve_worker_in_group` abstract stub so the
  wave-4 `_TaskOpsMixin.claim` cross-mixin caller resolves under
  pyright; the composed backend's MRO routes to `_GroupOpsMixin`'s
  real body. `_validate_registry_id` already lives on `_StoreCore`
  (wave 1).
- `_StoreBase(_WorkerOpsMixin, _GroupOpsMixin, _StoreCore)`; the
  remaining task/idea/variant/event/experiment ops stay inline.

Behavior-preserving: storage suite 373 passed / 185 skipped, full
pyright clean, complexity gate green (annotation still present).
